Revert stalker_neytral_hood_9.ogf to vanilla model
This commit reverts the stalker_neytral_hood_9.ogf model used by some neutral NPCs back to the original vanilla model. This model features a gas mask with disjointed lenses and an oxygen tank. Previously, a variant with only a gas mask and backpack was used for these NPCs, which is now reserved for the player character.